64-bit warnings fixes
git-svn-id: svn://messdev.no-ip.org/mess@650 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/formats/cassimg.c
@@ -67,7 +67,7 @@ static double map_double(double d, UINT64 low, UINT64 high, UINT64 value)
 
 static size_t waveform_bytes_per_sample(int waveform_flags)
 {
-	return 1 << ((waveform_flags & 0x06) / 2);
+	return (size_t) (1 << ((waveform_flags & 0x06) / 2));
 }
